src/main/java/com/mycompany/ductumejb/servico/MeetingService.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.mycompany.ductumejb.servico;

import com.mycompany.ductumejb.entidade.Meeting;
import java.util.List;
import javax.ejb.LocalBean;
import javax.ejb.Stateless;
import javax.ejb.TransactionAttribute;
import static javax.ejb.TransactionAttributeType.SUPPORTS;
import javax.validation.executable.ExecutableType;
import javax.validation.executable.ValidateOnExecution;

/**
 *
 * @author deecarneiro
 */
@Stateless
@LocalBean
@ValidateOnExecution(type = ExecutableType.ALL)
public class MeetingService extends Servico<Meeting> {

    @Override
    public Meeting criar() {
       return new Meeting();
    }

    @TransactionAttribute(SUPPORTS)
    public List<Meeting> consultarEntidades() {
       return consultarEntidades( new Object[] {}, Meeting.ALL_MEETINGS);
    }
 
}
